Remove callback arguments from most functions.
After some contemplation, I realized that given the underlying behavior of 'wait_for_rsp', that
synchronous and asynchronous calls play well with each other.  The pattern of considering an 'onlogon'
function as running in a sort of greenthread works well and leads to much more comprehensible code
for the average person.  While removing those arguments, the module contents were reorganized to
serve as an example of how more straightforward code written in this fashion may be.

Change-Id: I9c83660e4d3c68ad8ae0c016a3377b2a0e7ee6ec