Rename to pyghmi
It has been expressed as a concern that 'ipmi' is too generic a name.
Additionally, it is also the case that non-ipmi capability is likely
to be incorporated as it goes along (e.g. Enclosure management and
virtual media are frequently not IPMI based).

Move existing content under the 'pyghmi' namespace.  pyghmi stands for
'python general hardware management infrastructure' and is pronounced
'pygmy'

Change-Id: Ib549a9f5b7dd549c7dc5ddbab251a2e06c572e41