Enhance wait behavior of set_power
Previously, wait would always be indefinite and act in many ways like
a busy wait.  The check for completion now takes one second between
every check.  The default behavior now goes for 300 seconds, but
wait can now take an int to indicate a different preference by the
caller.  300 seconds may seem like a lot, but the 'softoff' request
in particular suggests waiting for OS shutdown process to complete.

Additionally, some bugs in the retry logic were noted in the process
of having the retry timer drive the delay and are corrected.

Change-Id: Ibe34e87c2a58f13981d60e5f80b4b636e67ac3f9