Change SOL code to not send empty data to caller
If a SOL target sends a packet with no new data (e.g. spurious retry), do not
bother the caller with such empty data.  It's a waste of their attention at
best and if the caller is doing something like trying to write the data
straight to a TLS socket, it will trigger EOF exception.

Change-Id: I1d751331bc0fa0087bf671de12507b7fe011f5fa