Remove callback from session.py
In the private session module, there was originally a thought
that a callback model was going to be the path for code to implement
interleaved execution.  However, since that point a coroutine pattern
has proven to afford the same benefits with simpler looking calling code.
With this in mind, remove some of the disused code in order to make
the session module a little easier to follow.

Change-Id: Ice697df409dd188d7e3361d278f8b1ce1358a419