Use different timeout values for logon and retry
When establishing a session, there is not much to lose.
If three get channel authentication capabilities fail, it's
almost certainly because it just isn't going to work (bad IP,
rebooting BMC, something like that).  However, once signs of life
have been displayed, the likely cause of an outage shifts toward
a transient network outage (e.g. link speed negotiation after
rebooting a shared nic configured solution).  Particularly with SOL,
there is a high chance of inflight data worth trying to recover the
session to get.  For that we go to tolerating up to ~20 seconds of
outage.  Much longer than that and we are pretty much certain to run
afoul of BMC timeout anyway, so no hope of withstanding something like
spanning tree with blocking by default, but it can withstand moderately
well configured networks.

Change-Id: I783d13eaee4360cfd780e28b2b4cbf77cc980945