Do not count empty SOL data as keepalive data
If the console layer sends an SOL payload with zero data (which is
valid to do) the session object would treat that as keeping the
session alive on the BMC side.  Many BMCs do not consider empty
SOL payload packets toward session activity.  Accomodate by having
the keepalive deferral not be triggered by such packets.

Change-Id: I8d278dbc48fae296c754ce9bc8f05bc92b560374