Gracefully flag unrecognized discrete sensor states
Previously, pyghmi would incur a KeyError in the event of
something like sensor specific event not in the lookup tables.
Fix this by catching the error and putting in generic description
of state and flagging as 'warning'.  This sort of warning suggests
a bug in pyghmi that pyghmi needs to have its event descriptions
updated.

Change-Id: Ia5f4645e30a341a93456dae2b8d75d9fda542e1b