Use multiple sockets instead of one
Given the nature of SOL, it is impossible to control flow of incoming
traffic.  This means measures to mitigate the risk of exhausting buffer
memory on the socket cease to be effective.  Modify strategy to stop
throttling and instead allocate new sockets to acquire more network
buffer space.  This means the footprint of a small scale setup is
actually even lower and a larger setup does get more filehandles,
but still 1/64th the footprint of the usual strategy.

Change-Id: I10698393d31b0c04d0242ff85815239078c076e2