Correct pool usage count on failed login
If trying to establish a session and an error occurs, pyghmi
was not decrementing the usage count on the pool, leading for
more than expected filehandles being created over time.  Fix
this by correctly decrementing the count in the case where
something is not yet broken, but is also not yet logged in.

Change-Id: I1b27f9b3b902a253d38293182305cc4dac26b765