Fix concurrent raw_command calls to Session
If two contexts call raw_command concurrently, there was a scenario
where the first to transmit has its result overwritten by the next to send
and corrupts the results of the first command.  One scenario where this was
encountered was when a get health call was being serviced at the same moment
SOL attempted to open a console, causing one of the get sensor readings to
complain that 'SOL was already active'.  Address it by storing away lastresponse
in a more context specific place before deasserting 'incommand' and remove
instances that deasserted it earlier.

Change-Id: I504da3f54562a4b65b8f4e9e20c19aed9d21a09f