Remove overly aggressive packet processing
With the socket pooling, we no longer have to be as aggressive about trying to get packets out of the socket
buffer.  However, in a very busy system of several hundred nodes, this aggressive processing tends to produce
larger python stacks and can exceed the default 1,000 limit.  We may have to resort to increasing the limit one
day, but for now this case can be avoided.

Change-Id: I83bdaa2d8ad464727a69e6ece064f51cd4318822