Rework IO Worker thread behavior
python interpreter exit is very chaotic with a daemon thread.
Address this by making the worker thread non daemon, and
hooking join() to do the exit instead.  Do a particular
dance to accomodate a caller that may replace python's library
after we are imported by defining the class at the same time
the threading library was originally referenced.

Change-Id: I8ebd2d4e89b4e11e352e440775fd236599c024a0