Correct delay_xmit behavior
delay_xmit was broken due to deferral of the retry calculation.
It relied upon the deadline set in the retry to govern the
initial transmit of data.  By trying to make it delay
to a more realistic time by not starting the clock
until the packet was transmitted, the delay_xmit
which never transmitted was in bad shape.  This was already
broken since retry=False would have had the same effect, but
that is a rare scenario.  Fix by having delay_xmit specifically
handle the assignment separate of retry logic.

Change-Id: Ia251215f14f8a5808e8a8f5b3f53fbef40721709