Handle concurrent session requests
A caller may end up making requests that would get gathered
into a single session object.  However due to being aggressive,
a second session object supersedes the original without the
original getting a chance to complete login and satisfy the
second caller.  Address this by reusing a 'logging' session
and then having __init__ follow the state of the other session
object already in progress.  Without this a caller can end
up provoking a fight and having the BMC refuse to continue
to entertain the shenanigans in short order.

Change-Id: I47acbc0c974900ff50c02d470b5a79a3fed9bb73