Fix needless retries due to misdirected packets
When used in a threaded application, there was a chance for one wait
iteration to slurp away packets from another instance.  When this
would happen, a wait may mistakenly act as if no packets were
received and expire a session timeout. Fix this by having arbitrarily
many instances feed and consume from the same queue.  This way if any
instance of the wait function pulls a packet, all consumers are made
aware of the packet. This dramatically improves performance when
dealing with very long conversations (SDR+sensors) with hundreds of
nodes across hundreds of threads.

Change-Id: I3f51097fb41197445a447cbdaddc8c1c29d4a873