Fix exceptions on sdr read
The string formatting is corrected for some 'bmc' errors.
It has been suggested to restructure things to not complain,
but so far when investigated, the systems actually had
a defect in putting the wrong sensor number in.  As a
diagnostic aid, this seems to be useful.  If a legitimate
application of duplicate SDR record for sensor is pointed
out, then we can restructure.

Change-Id: I58c2ffc1108cbb157f1398b420ea3a24bc4f05e8