Add system UUID to inventory
While not strictly in the FRU area, it is often desirable
to have the system UUID available.  The intent is for the
UUID to match what dmidecode would return.  If a manufacturer
does it right, that UUID will be unique.  For ThinkServers,
override with the UUID from the OEM FRU fields rather than
using the get system UUID result.

Change-Id: Ie9a1b7e8fee2cb40ab679cbf2df04db61fd4e42f