Rename description field and split data out
It would be good for some consumers to have the broad event
description separated from the event specific data.  Facilitate
this with language that makes more clear the piece that refers
broadly to a class of things that can go wrong versus the
potentially very specific data associated with that event.

Change-Id: I983dc1cdf59ae8bc08db3ff3f55192e82f693dbe