Report/Configure LAN alert destinations
Provide a facility to manipulate the LAN alert destinations.
Additionally update the LAN configuration functions to have
a function to intelligently find the correct LAN channel.
For now the 'check current channel' preference should always
activate, but the function should be able to cover a hypothetical
in-band scenario should session become so enabled.

Change-Id: I45ce81addb882110e04d0526997a96b6b2ab0c2e