Have session.py work better in Windows
Have monotonic_time reworked to use Windows function
when available.  Eliminate use of an os.pipe to break
into the select.  For the most part, move to using
the threading.Event wait(timeout) scheme rather than
breaking into the select.  For the part where we want to
interrupt the select, use the first udp socket as the
angle to break in.  Additionally, provide a Windows
friendly value for IPPROTO_IPV6 should the socket module
fail us.

Change-Id: I8f566543c8da28eb7ed76f1cb80ff4cb2dcbba96