Extend and document user_delete in ipmi.command
User deletion has some limitations making the function a best effort.
Document this and provide at least one additional attempt at deleting
a user to cover more vendors.  This still does not cover all vendors,
and OEM extensibility is likely to be the answer to allow this to
cover more implementations.

Change-Id: I6af3f761c587f9eb8b58b3408052d6585847f7d6