Workaround likely spec deviations and fix parsing
For one, was doing UTF-8, though the encoding strictly speaking
per speac should be ISO-8859-1.  Corrected that mistake.  Also
workaround an issue when FRU may be programmed with 0x0 or 0xff
bytes to terminate, or spaces.  The spec actually doesn't speak
to a strategy to pad strings (presumably because they assumed
implementation would avail itself of the capacity for variable
strings and not need to deal with silly fixed length fields).
Handle this by trimming out trailing spaces, 0x0, or 0xff.  This
strictly speaking violates the spec if they want to use one or more
ÿ to end a field, but will take the risk that no one actually would
want to do that.

Change-Id: I7d7e3aa9f3fd2b1e1af75f98cabd49ca374755c0