Avoid abandoning retry when SOL packet is not ACK
The design was to abandon retry when data comes in,
ACK or not.  This became a problem when console.py
retried under frantic conditions, confusing some
BMC implementations, and so that was removed.
With that removed, BMC traffic was still cancelling
the retry, but nothing was reinstated in retry anymore.
Fix by having session.py understand a few bits of SOL
payload enough to identify at least if it is *an* ACK
at all.  If it's one of the 'weird' ACK scenarios,
sol handler still must handle, but simple no-ACK scenario
now contained in session.py.

Change-Id: I609ff783cd6cc2bf431271280bb55da4126f2cfc