Avoid stall waiting on incommand
It was possible for multiple command requests to be pending.  While pending,
if a waiter was awoken first, *then* the satisfied command, then the waiter
will stall in the next iteration, recovering only after the failsafe mechanism kicks
in.  Address by having the instance clearing incommand set() the next event in
line.

Change-Id: I8dbd08027596b4acccace6599f996ad547a47768