Implement M4 and Kent support
M4 and Kent are two IBM systems - this explain why there is a new oem
called 'ibm') that are under Lenovo's responsibility.
This implementation was coordinated by internal Lenovo development team
in China (Michael Du as team leader) and kept under control in an
internal fork that is being gradually closed.

Change-Id: I244a9e9f3a7466b01e429eeba47caec60780496f
Signed-off-by: Fabio Dassan <fdassan@lenovo.com>