IpmiServer to allow using IP version 4
This patch is introducing a new parameter for the IpmiServer called
"use_ipv4", when set to True it allows the use of an IP version 4 for
the server. By default it's set to False (IPv6 only).

This patch also fixes a problem when binding a IPv4 address, in the IPMI
Session class. Prior to this patch it would fail with: "Error:
getsockaddrarg() takes exactly 2 arguments (4 given)"

Change-Id: I802c4e2dbe140da40f3f3111aa84d692d4374bb1