Localize console lock
The console code was using a global lock, but the data protected
is per instance.  Change to using a local version.  I couldn't figure out
a generic way for the decorator approach to dereference 'self'
without creating a specific requirement on the class name, so
I changed to using it as a context manager directly for now.

Change-Id: I304b75572fecd3326723d981fec67207d87d0742