Replace waitall mechanism with never wait
Previously waitall was used as a way to have wait_for_rsp be able to
call out for specifically keepalive, to prevent a wait_for_rsp from
hanging.  The thought was to create a new wait_for_rsp by recursing.
However, if things go really bad with enough systems, it runs afoul of
python's recursion limit.  The alternative strategy is to provide a way
to never wait, and asynchronously do cleanup in a callback.  This
returns to the original wait_for_rsp instance rather than recursing
to a new one, keeping the stack shallow.

Change-Id: I54450bfc5bf57563a0a8bd986d9aea887bf36e53