Prefer cyrptodomex if present
For compatibility, accept PyCrypto.  However if Cryptodomex is present
(explicit version of cryptodome), use that.

This is an easier path to a modern supported AES implementation that
continues to support older, but still prevalent long term linux
distributions.

Another project required cryptodome, and evidently it caused problems
due to some interoperability with things that truly needed PyCrypto.

Change-Id: I3c0601bc0b6ae56417e712b50337ec1c20eb89a2