Provide a grace period for session validity
If a session has *just* expired, it may just be waiting for the
next timeout pass.  If it expired 15 seconds ago, however, it should
by all rights be beyond dead.

This matters as the calling code may be oblivious to the change of
session object under the covers if this is tripped.

Change-Id: I5be3bafb5932c287eb17c66217ba630746533310