Fix break handling in SOL
When the retry logic was moved to console, it exposed a problem with
the sol handler being called in the midst of transmitting without
clearing the pending payload, resulting in an infinite recursion
of sending the same break.

Resolve by using the common call to sendoutput to have it occur
only after the current head of pending output has been removed.

Change-Id: Iaae86a3bd870c1841530e200ef0f5780a5cc6993