Genericize bulk requests and fallback to Id
Independent concurrent requests are going to be a way of life, add
a generic helper.  Future modify to take a dict of urls to handlers
to launch handlers in a pool as well.

Names are usually more descriptive, but are occasionally not unique.
Fallback to Id for devices that conflict.

Store name map to urls for faster performance of
get_inventory_of_component

Change-Id: I83207d24c8bd2a23ac0d16ab5e1777f7df8d4369